+++ pc/src/main/java/com/allen/service/basic/producelinecoreproduct/AddProduceLineCoreProductService.java
@@ -4,5 +4,5 @@ package com.allen.service.basic.producelinecoreproduct;
  * Created by Allen on 2016/12/22 0022.
  */
 public interface AddProduceLineCoreProductService {
-    public void add(long plId, long wcId, String delPlcpIds, Long[] plcpIds, Long[] pIds, Integer[] qualifiedRates, String loginName)throws Exception;
+    public void add(long plId, long wcId, String delPlcpIds, Long[] plcpIds, Long[] pIds, Float[] qualifiedRates, String loginName)throws Exception;
 }
